Show builder error messages in hover
Error messages from the typechecker and from the builder need to be
handled differently. The error messages from the builder are not
available as long as the user doesn't build the project, whereas
the presentation compiler can show error messages immediately.

Therefore, error messages from the builder are more expensive to show.
We first have to check if the PC finds error messages and only if it
doesn't return anything, we should continue with checking if the builder
reported anything.

The error messages from the builder are available through markers.
Because they need at least be provided on a single line (because the
problems view can't handle multi line strings) the error message is
saved two times - a single line message for the problems view and the
full error massage for the hover.

Fixes #1001534